app/backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-speech
python-multipart
